Updated Interface
Completely changed the way the player interacts with candy. Instead of
running over candy, the player looks at the candy with the reticle and
the candy will be highlighted. Clicking the candy will destroy the candy
and increase the score.

* Removed Script_Player from the Player Controller. We don't need it
with this new way of interacting with candy. The script is still there
if we need to use it later.
* Added a bunch of stuff to Script_GUI. Everything is being handled here
now.
* Increased the the time limit to 300 seconds to make it easier to test.