Ghost and Behavior
Added a Finite State Machine that handles Ghost AI. There are 4 states:
1. Idle
2. Search
3. Chase
4. Die
In 'Idle' the ghost hovers up and down at it's spawn position. In
'Search' it picks a random position and moves to it. In 'Chase' it
chases after the player. In 'Die' it destroys itself.

The ghost uses a Vector3.Angle and compares that to it's field of view
and viewing distance to check if it can see the player. Since it's a
ghost it doesn't check for any obstacles, I'm allowing it to see and
move through walls.