Game Over when all candies are grabbed.
Changed it so that the game is over when the player gets all the candy.
Added comments and changed the timer to help testing the new win
condition.
Changed the collider on the M&Ms from a Mesh collider to a Sphere
collider to make it easier to pick them up.