Ghost Scare state
Added new ghost state. When the ghost is close to the player it enters
the Scare state. This triggers behavior meant to scare the player. Right
now it jerks left and right really fast in a kinda freaky manner. I want
it to make the lights pulse bright and dark as it gets close.