Search Targets
Updated the ghost Search behavior to use Search Targets, empty game
objects I added to the scene to act as nodes. When the ghost enters the
Search state it picks a random node to move to. After it reaches the
node it picks another, and so one for 4 tries. If it doesn't find the
player in 4 tries the ghost disappears and another ghost will spawn at a
random time.