Dropped Candy
* Added a new mechanic to the Scare state. When the ghost is scaring the
player it drops the player's score by 10 once every second. This is to
simulate the player dropping candy in fright.
* Updated the FSM to transition from Scare back to Chase. Previously the
ghost would always stay in the Scare state until it couldn't see the
player anymore and then go straight to the Search state. Now once the
player backs up to more than 4 points away the ghost back to the Chase
state. It feels like a more natural action.