Changed font
* Added the 'Chiller' font to the game and implemented it in the GUI.
Increased the size of the text in all GUI elements so its easier to
read.
* Added the 'DestroyCandy' function from CandySpray to every state
transition out of Scare to fix a bug where the sprayed candy would stay
in the world and roll around if the player ran away when the candy still
existed.
* Added a time bonus to the 'win' game over screen. I add the time left
to the current score. This way the faster the player completes the level
the higher their score will be.